+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release date: ??.??.????
 * IE11: Table of contents visual fix (HUB-328)
 * Deprecated code cleanup (HUB-337)
 * Better search results (HUB-260)
+* Sort degree programmes in a pre-defined order (HUB-336)
 
 
 ## 1.20